+++ Scripts/Editor/Database/Commands/SetViewIdCommand.cs
@@ -1,6 +1,6 @@
 using System;
-using System.Linq;
 using TeamODD.ODDB.Runtime.Interfaces;
+using TeamODD.ODDB.Runtime.Mutations;
 using TeamODD.ODDB.Runtime.Utils.Converters;
 
 namespace TeamODD.ODDB.Editors.Commands
@@ -11,8 +11,6 @@ namespace TeamODD.ODDB.Editors.Commands
         private readonly ODDBID _oldId;
         private readonly ODDBID _newId;
         private readonly Action<string> _notifyChanged;
-        private IView _view;
-        private int _repositoryIndex = -1;
         private bool _captured;
 
         public override string Name => "Set View ID";
@@ -41,7 +39,7 @@ namespace TeamODD.ODDB.Editors.Commands
 
         public override void Undo()
         {
-            if (!_captured || _view == null || SameId(_oldId, _newId))
+            if (!_captured || SameId(_oldId, _newId))
                 return;
 
             ReKey(_newId, _oldId);
@@ -53,40 +51,12 @@ namespace TeamODD.ODDB.Editors.Commands
             if (_captured)
                 return;
 
-            _view = _repository.Read(_oldId);
-            _repositoryIndex = IndexOf(_view);
-            _captured = true;
+            _captured = _repository.Read(_oldId) != null;
         }
 
         private void ReKey(ODDBID fromId, ODDBID toId)
         {
-            var view = _repository.Read(fromId);
-            if (view == null)
-                return;
-
-            _repository.Delete(fromId);
-            view.ID = toId;
-            _repository.Update(toId, view);
-            RestoreRepositoryIndex(view);
-        }
-
-        private void RestoreRepositoryIndex(IView view)
-        {
-            if (_repositoryIndex < 0)
-                return;
-
-            var currentIndex = IndexOf(view);
-            if (currentIndex < 0 || currentIndex == _repositoryIndex)
-                return;
-
-            _repository.Move(currentIndex, _repositoryIndex);
-        }
-
-        private int IndexOf(IView view)
-        {
-            if (view == null)
-                return -1;
-            return _repository.GetAll().ToList().IndexOf(view);
+            ODDBMutations.RekeyRepositoryItem(_repository, fromId, toId);
         }
 
         private void Notify(ODDBID staleId, ODDBID activeId)